https://github.com/umjammer/vavi-util-archive-sandbox/security/code-scanning/2?query=ref%3Arefs%2Fheads%2Fmaster

+++ src/main/java/vavi/util/rar/UnRar.java
@@ -1261,7 +1261,7 @@ public class UnRar {
         while (count-- > 0) {
             pn2 += pn3;
             pn1 += pn2;
-            data[p] += pn1;
+            data[p] += (byte) pn1;
             p++;
         }
     }
